# docs/swagger.yaml
basePath: /api
definitions:
  authentication.LDAPEndpointType:
    properties:
      domain_names:
        items:
          type: string
        type: array
      hostname:
        type: string
      port:
        type: integer
      single_domain:
        type: boolean
      tls_enable:
        type: boolean
    type: object
  authentication.RadiusEndpointType:
    properties:
      hostname:
        type: string
      port:
        type: integer
      secret:
        type: string
    type: object
  gin.H:
    additionalProperties:
      type: any
    type: object
  types.AuthorizedResponseType:
    properties:
      issue:
        type: string
      redirect_url:
        type: string
      status:
        type: string
    type: object
  types.CaptivePortalConfigFundamentalType:
    properties:
      domain_names:
        items:
          type: string
        type: array
      html:
        $ref: '#/definitions/types.HTMLType'
      mode:
        type: string
      single_domain:
        type: boolean
    type: object
  types.CheckCredentialType:
    properties:
      password:
        type: string
      secret:
        type: string
      username:
        type: string
    type: object
  types.ConfigType:
    properties:
      administrator:
        $ref: '#/definitions/types.CredentialType'
      allow_endpoints:
        items:
          $ref: '#/definitions/types.EndpointType'
        type: array
      bypass_networks:
        items:
          type: string
        type: array
      ddos_prevention:
        type: boolean
      domain_names:
        properties:
          auth_domain_name:
            type: string
          operator_domain_name:
            type: string
        type: object
      egress_interface:
        type: string
      external_portal_url:
        type: string
      fqdn_blocklist:
        items:
          type: string
        type: array
      html:
        $ref: '#/definitions/types.HTMLType'
      ldap:
        $ref: '#/definitions/authentication.LDAPEndpointType'
      max_concurrent_session:
        type: integer
      radius:
        $ref: '#/definitions/authentication.RadiusEndpointType'
      redirect_url:
        type: string
      secure_interface:
        type: string
      session_idle:
        type: integer
    type: object
  types.CredentialType:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  types.EndpointType:
    properties:
      hostname:
        type: string
      port:
        type: integer
    type: object
  types.ExtendConfigType:
    properties:
      administrator:
        $ref: '#/definitions/types.CredentialType'
      allow_endpoints:
        items:
          $ref: '#/definitions/types.EndpointType'
        type: array
      bypass_networks:
        items:
          type: string
        type: array
      ddos_prevention:
        type: boolean
      domain_names:
        properties:
          auth_domain_name:
            type: string
          operator_domain_name:
            type: string
        type: object
      egress_interface:
        type: string
      external_portal_url:
        type: string
      fqdn_blocklist:
        items:
          type: string
        type: array
      html:
        $ref: '#/definitions/types.HTMLType'
      ldap:
        $ref: '#/definitions/authentication.LDAPEndpointType'
      max_concurrent_session:
        type: integer
      radius:
        $ref: '#/definitions/authentication.RadiusEndpointType'
      redirect_url:
        type: string
      secure_interface:
        type: string
      session_idle:
        type: integer
      status:
        properties:
          egress_ip_address:
            type: string
          secure_ip_address:
            type: string
        type: object
    type: object
  types.HTMLType:
    properties:
      background_file_name:
        type: string
      default_language:
        type: string
      en_sub_title:
        type: string
      en_title_name:
        type: string
      logo_file_name:
        type: string
      th_sub_title:
        type: string
      th_title_name:
        type: string
    type: object
  types.InitializedType:
    properties:
      ip_address:
        type: string
      secret:
        type: string
    type: object
  types.SessionType:
    properties:
      ip_address:
        type: string
      issue:
        type: string
      last_seen:
        type: string
      session_uuid:
        type: string
    type: object
info:
  contact: {}
  description: This is a COCO Captive Portal
  license:
    name: Apache License Version 2.0
    url: https://github.com/mrzack99s/coco-captive-portal
  title: COCO Captive Portal
  version: "1"
paths:
  /api/adm-signed:
    get:
      consumes:
      - application/json
      description: Get admin signed
      operationId: adm-signed
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get admin signed
      tags:
      - Authentication
  /api/authentication:
    post:
      consumes:
      - application/json
      description: Check credential to get access
      operationId: authentication
      parameters:
      - description: Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/types.CheckCredentialType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.AuthorizedResponseType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Authentication
      tags:
      - Authentication
  /api/check-is-administrator:
    post:
      consumes:
      - application/json
      description: Check is administrator
      operationId: check-is-administrator
      parameters:
      - description: Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/types.CredentialType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Check is administrator
      tags:
      - Operator
  /api/config:
    get:
      consumes:
      - application/json
      description: Get config
      operationId: get-config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ExtendConfigType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Get config
      tags:
      - Operator
    put:
      consumes:
      - application/json
      description: Set config
      operationId: set-config
      parameters:
      - description: Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/types.ConfigType'
      produces:
      - application/json
      responses:
        "200":
          description: updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Set config
      tags:
      - Operator
  /api/count-all-session:
    get:
      consumes:
      - application/json
      description: Count all session
      operationId: count-all-session
      produces:
      - application/json
      responses:
        "200":
          description: count
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Count all session
      tags:
      - Operator
  /api/get-all-session:
    get:
      consumes:
      - application/json
      description: Get all session
      operationId: get-all-session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.SessionType'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Get all session
      tags:
      - Operator
  /api/get-captive-portal-config-fundamental:
    get:
      consumes:
      - application/json
      description: Captive portal config fundamental
      operationId: get-captive-portal-config-fundamental
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.CaptivePortalConfigFundamentalType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Captive portal config fundamental
      tags:
      - Utils
  /api/initialize:
    get:
      consumes:
      - application/json
      description: Get initialize secret to get access
      operationId: initialize
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.InitializedType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get initialize secret to get access
      tags:
      - Authentication
  /api/is-exist-initialize-secret:
    post:
      consumes:
      - application/json
      description: Exist initialize secret
      operationId: is-exist-initialize-secret
      parameters:
      - description: Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/types.InitializedType'
      produces:
      - application/json
      responses:
        "200":
          description: found
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Exist initialize secret
      tags:
      - Authentication
  /api/kick-ip-address:
    put:
      consumes:
      - application/json
      description: To kick via ip address
      operationId: kick-via-ip-address
      parameters:
      - description: Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/types.SessionType'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: To kick via ip address
      tags:
      - Operator
  /api/kick-username:
    put:
      consumes:
      - application/json
      description: To kick via username
      operationId: kick-via-username
      parameters:
      - description: Parameters
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/types.SessionType'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: To kick via username
      tags:
      - Operator
  /api/net-intf-usage:
    get:
      consumes:
      - application/json
      description: Network Interfaces Usage
      operationId: net-interfaces-bytes-usage
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Network Interfaces Usage
      tags:
      - Operator
  /api/revoke-administrator:
    get:
      consumes:
      - application/json
      description: Revoke administrator
      operationId: revoke-administrator
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - ApiKeyAuth: []
      summary: Revoke administrator
      tags:
      - Operator
  /api/sign-out:
    get:
      consumes:
      - application/json
      description: Sign out
      operationId: sign-out
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Sign out
      tags:
      - Authentication
  /api/signed:
    get:
      consumes:
      - application/json
      description: Get signed by ip address
      operationId: signed
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SessionType'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get signed by ip address
      tags:
      - Authentication
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: api-token
    type: apiKey
swagger: "2.0"